- DB를 통한 Component 불러오기 및 초기화 가능 구현

+++ ApplicationBase/AInstancedModelComponent.h
@@ -5,7 +5,7 @@
 class AInstancedModelComponent : public AComponent
 {
 public:
-	AInstancedModelComponent(const uint64_t& componentID, DirectX::XMFLOAT3 position);
+	AInstancedModelComponent(const uint32_t& componentID, DirectX::XMFLOAT3 position);
 
 protected:
 	std::vector<ModelMaterialAsset*> m_selectedMaterials;
